change printed representation
please modify this code so that red kittens are printed as 'r', red
cats as 'R', blue kittens as 'b', and blue cats as 'B'

--

The main changes are:
In place_piece, instead of using f"{color}_kitten" and f"{color}_cat",
we use f"{color[0]}" to represent a kitten of that color.
In push_piece, we split self.grid[row][col] into color and piece and
assume that piece is 'kitten'.
In has_three_in_row, we check for the uppercase letter for cats and
the lowercase letter for kittens.
In has_eight_on_board, we check for the lowercase letter of the color.
In upgrade_three_kittens and upgrade_single_kitten, we use the
lowercase letter for kittens and the uppercase letter for cats.
Now, the board will print 'r' for red kittens, 'R' for red cats, 'b'
for blue kittens, and 'B' for blue cats.