only push one space
When an adjacent piece is pushed, it is only pushed one space. Please
fix the code.

--

The only change is in the push_piece method:
Now, when pushing a piece, it will only move one space in the given
direction if the destination square is empty and on the board. If the
piece cannot be pushed, it is removed from the board and added back to
the respective player's supply.
With this change, adjacent pieces will only be pushed one square away
from the newly placed piece, as per the game rules.